feat(server): object create side effect self contained view creation
